First pass at multiple author paths.
Some minor bugs fixes for the errors and added packages needed to get
running.
On CNN.com not all authors will have a link. So I change the
authorSelector to be an array that has multiple selector paths.
Started work on a authorCleanup method,  but the getting ‘TypeError:
site.authorCleanup is not a function’. Not sure why the function is not
being exported.